+++ dst/app/middlewares/authentication.js
@@ -57,16 +57,15 @@ exports.default = (req, res, next) => __awaiter(void 0, void 0, void 0, function
                 if (user.username !== undefined) {
                     programMembership = {
                         membershipNumber: user.username,
-                        programName: { en: 'Amazon Cognito', ja: 'Amazon Cognito' },
-                        project: req.project,
-                        typeOf: cinerinoapi.factory.programMembership.ProgramMembershipType.ProgramMembership,
+                        project: { typeOf: req.project.typeOf, id: req.project.id },
+                        typeOf: cinerinoapi.factory.chevre.programMembership.ProgramMembershipType.ProgramMembership,
                         url: user.iss
                     };
                 }
                 req.user = user;
                 req.accessToken = token;
                 req.agent = {
-                    typeOf: ttts.factory.personType.Person,
+                    typeOf: cinerinoapi.factory.personType.Person,
                     id: user.sub,
                     memberOf: programMembership,
                     identifier: identifier