+++ dst/app/routes/reservations.js
@@ -355,7 +355,7 @@ function cancelReservation(params) {
             project: project,
             typeOf: ttts.factory.chevre.transactionType.CancelReservation,
             agent: {
-                typeOf: ttts.factory.personType.Person,
+                typeOf: cinerinoapi.factory.personType.Person,
                 id: 'tokyotower',
                 name: '@tokyotower/domain'
             },